Two significant changes.
1. dbChan is now channel of []fsentry.E, not single ones
   This should prevent the dbChan from filling up.
2. There are now DELETE PreparedStmts - one to delete an individual item
   and one to delete children - the latter is used when a directory is deleted.
Updated the unit tests to match.

There is a flaw in the overall design - if a new file or directory is
added to a child directory that is not top-level, it will not be seen
by fslocate indexer, since it won't descend into dirs that is has already
seen (i.e., are in the db).  Redesign required.